backend/requirements.txt
anyio==3.6.2
CacheControl==0.12.11    
cachetools==5.3.0        
certifi==2022.12.7       
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1     
Deprecated==1.2.13       
distlib==0.3.6
fastapi==0.95.0
filelock==3.10.7
firebase==4.0.1
firebase-admin==6.1.0
Flask==2.2.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.83.0
google-auth==2.17.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.10.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.53.0
gunicorn==20.1.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jwcrypto==1.4.2
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.2
oauth2client==4.1.3
pandas==1.5.3
platformdirs==3.2.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
PyJWT==2.6.0
pyparsing==3.0.9
Pyrebase4==4.6.0
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
starlette==0.26.1
threadpoolctl==3.1.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.21.0
Werkzeug==2.2.3
wrapt==1.15.0